+++ src/main/java/com/owlmaddie/particle/ParticleEmitter.java
@@ -1,13 +1,18 @@
 package com.owlmaddie.particle;
 
 import net.minecraft.entity.Entity;
-import net.minecraft.server.world.ServerWorld;
-import net.minecraft.util.math.MathHelper;
 import net.minecraft.particle.DefaultParticleType;
+import net.minecraft.server.world.ServerWorld;
+import net.minecraft.sound.SoundCategory;
+import net.minecraft.sound.SoundEvents;
+import net.minecraft.util.math.MathHelper;
+
+import static com.owlmaddie.network.ServerPackets.FIRE_BIG_PARTICLE;
+import static com.owlmaddie.network.ServerPackets.HEART_BIG_PARTICLE;
 
 public class ParticleEmitter {
 
-    public static void emitCreatureParticle(ServerWorld world, Entity entity, DefaultParticleType particleType) {
+    public static void emitCreatureParticle(ServerWorld world, Entity entity, DefaultParticleType particleType, double spawnSize, int count) {
         // Calculate the offset for the particle to appear above and in front of the entity
         float yaw = entity.getHeadYaw();
         double offsetX = -MathHelper.sin(yaw * ((float) Math.PI / 180F)) * 0.9;
@@ -20,6 +25,13 @@ public class ParticleEmitter {
         double z = entity.getZ() + offsetZ;
 
         // Emit the custom particle on the server
-        world.spawnParticles(particleType, x, y, z, 1, 0, 0, 0, 0);
+        world.spawnParticles(particleType, x, y, z, count, spawnSize, spawnSize, spawnSize, 0.1F);
+
+        // Play sound when lots of hearts are emitted
+        if (particleType.equals(HEART_BIG_PARTICLE) && count > 1) {
+            world.playSound(entity, entity.getBlockPos(), SoundEvents.ENTITY_EXPERIENCE_ORB_PICKUP, SoundCategory.PLAYERS, 0.4F, 1.0F);
+        } else if (particleType.equals(FIRE_BIG_PARTICLE) && count > 1) {
+            world.playSound(entity, entity.getBlockPos(), SoundEvents.ITEM_AXE_STRIP, SoundCategory.PLAYERS, 0.8F, 1.0F);
+        }
     }
 }